clean up + minimal doc
