--- slave_stack/examples/nuttx/README.md
@@ -1,1 +1,0 @@
-TODO :)
